Move `ThumbnailMixin` out to `mixins.py` and update documentation accordingly
